initial setup for eip-1559 changes
