+++ chain/bsc/tx.go
@@ -1,16 +1,16 @@
 package bsc
 
 import (
-	"github.com/renproject/multichain/chain/ethereum"
+	"github.com/renproject/multichain/chain/evm"
 )
 
 type (
-	// TxBuilder re-exports ethereum.TxBuilder.
-	TxBuilder = ethereum.TxBuilder
+	// TxBuilder re-exports evm.TxBuilder.
+	TxBuilder = evm.TxBuilder
 
-	// Tx re-exports ethereum.Tx.
-	Tx = ethereum.Tx
+	// Tx re-exports evm.Tx.
+	Tx = evm.Tx
 )
 
-// NewTxBuilder re-exports ethereum.NewTxBuilder.
-var NewTxBuilder = ethereum.NewTxBuilder
+// NewTxBuilder re-exports evm.NewTxBuilder.
+var NewTxBuilder = evm.NewTxBuilder
