+++ chain/evm/encode_test.go
@@ -1,12 +1,12 @@
-package ethereum_test
+package evm_test
 
 import (
 	"encoding/hex"
 	"fmt"
-	"github.com/renproject/multichain/chain/ethereum"
 	"math"
 	"testing/quick"
 
+	"github.com/renproject/multichain/chain/ethereum"
 	"github.com/renproject/pack"
 
 	. "github.com/onsi/ginkgo"
